Simplify shim structure, build from Rust source
The shim is now two parts, code and data. See code comments for details.

Also we build the shim from Rust code. Unfortunately some assembly code
is still required as the ELF entrypoint does not work like a C function,
but it should be much more readable and portable than the original blob
of bytes.

To avoid having to ask the user to install more targets, the shim is
written as a `#![no_core]` executable. Fingers crossed this never gets
unfixably broken.